fix: update config handling in madrat archive copy script
